requirements_txt.txt
Flask==3.0.0
flask-cors==4.0.0
mysql-connector-python==8.2.0
gunicorn==21.2.0